update: upload all form images at once during form creation
