+++ amplify/.config/local-env-info.json
@@ -1,5 +1,5 @@
 {
-  "projectPath": "/Users/Micah/projects/VALIDATEAlpha",
+  "projectPath": "/Users/Admin/Desktop/Epicodus/VALIDATE/VALIDATEAlpha",
   "defaultEditor": "vscode",
   "envName": "feat"
 }
